pyproject.toml
# Configuración del proyecto para uv
[project]
name = "tarracograf"
version = "0.1.0"
description = "Plataforma ciudadana para iniciativas cívicas en Tarragona"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Flask==3.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Security-Too==5.3.3",
    "Flask-Login==0.6.3",
    "Flask-WTF==1.2.1",
    "Flask-Mail==0.9.1",
    "Flask-Migrate==4.0.5",
    "Flask-Babel==4.0.0",
    "python-dotenv==1.0.0",
    "email-validator>=2.1.0",
    "Pillow==11.1.0",
    "gunicorn==21.2.0",
    "bcrypt==4.1.2",
    "bleach==6.1.0",
    "psycopg2-binary>=2.9.9",
    "setuptools>=65.0.0",
    "stripe>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

